refactor: メッセージ処理のリファクタリングなど
