+++ src/main/java/com/jaoafa/jdavcspeaker/Lib/EmojiWrapper.java
@@ -6,6 +6,7 @@ import com.vdurmont.emoji.EmojiParser;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
+import java.util.Objects;
 
 public class EmojiWrapper {
     /**
@@ -17,7 +18,7 @@ public class EmojiWrapper {
      */
     public static @NotNull String parseToAliases(@NotNull String input) {
         List<String> rawEmojis = EmojiParser.extractEmojis(input);
-        for (Emoji emoji : rawEmojis.stream().map(EmojiManager::getByUnicode).toList()) {
+        for (Emoji emoji : rawEmojis.stream().map(EmojiManager::getByUnicode).filter(Objects::nonNull).toList()) {
             String alias = ":" + emoji.getAliases().get(0) + ":";
             input = input.replaceAll(emoji.getUnicode(), alias);
             input = input.replaceAll(emoji.getTrimmedUnicode(), alias);
@@ -34,7 +35,7 @@ public class EmojiWrapper {
      */
     public static @NotNull String removeAllEmojis(@NotNull String input) {
         List<String> rawEmojis = EmojiParser.extractEmojis(input);
-        for (Emoji emoji : rawEmojis.stream().map(EmojiManager::getByUnicode).toList()) {
+        for (Emoji emoji : rawEmojis.stream().map(EmojiManager::getByUnicode).filter(Objects::nonNull).toList()) {
             input = input.replaceAll(emoji.getUnicode(), "");
             input = input.replaceAll(emoji.getTrimmedUnicode(), "");
         }
