+++ src/main/java/com/jaoafa/jdavcspeaker/Lib/LibValue.java
@@ -3,12 +3,7 @@ package com.jaoafa.jdavcspeaker.Lib;
 import com.rollbar.notifier.Rollbar;
 import net.dv8tion.jda.api.JDA;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class LibValue {
-    public static final List<String> ignoreContains = new ArrayList<>();
-    public static final List<String> ignoreEquals = new ArrayList<>();
     public static JDA jda;
     public static Rollbar rollbar = null;
 }
